requirements.txt
requests==2.31.0
apache-airflow-providers-http==4.8.0
